portal-api/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
psycopg==3.2.1
python-dotenv==1.0.1
python-jose==3.3.0
requests==2.32.3
